FastSerializer & Hashes have been moved from core_headers #403
