+++ sources/core/Hashes.h
