+++ sources/core/objects/Object.h
@@ -2,7 +2,7 @@
 
 #include "core_headers/KvAbort.h"
 #include "core/KvGlobals.h"
-#include "core_headers/Hashes.h"
+#include "core/Hashes.h"
 
 #include "../FastIsType.h"
 #include "../Idptr.h"
