+++ tests/saveable_operators_tests.cpp
@@ -1,6 +1,6 @@
 #include <gtest/gtest.h>
 
-#include "core_headers/Hashes.h"
+#include "core/Hashes.h"
 #include "core/SaveableOperators.h"
 
 TEST(SaveableOperators, MapOperatorsEmpty)
